# docs/02-api_contract.yaml
openapi: 3.0.3
info:
  title: chat2pay API
  version: 0.3.0
  description: |
    Reference API contract for the chat2pay POC.

    This file is kept for human alignment between frontend and backend.
    It is not intended to drive code generation.

    V1 scope:
    - profile selection and fixed POC access login
    - session and message history
    - registered payee lookup
    - domestic payment to a registered payee
    - optional streamed assistant responses from the Java backend
servers:
  - url: /api
    description: Relative API base path

tags:
  - name: Profiles
  - name: Auth
  - name: Chat

paths:
  /profiles:
    get:
      tags: [Profiles]
      summary: List available demo profiles
      operationId: listProfiles
      responses:
        '200':
          description: Available profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileSummary'

  /auth/profile-login:
    post:
      tags: [Auth]
      summary: Select a profile and validate the fixed POC access password
      operationId: profileLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileLoginRequest'
      responses:
        '200':
          description: Logged-in profile context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserContext'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout the current profile session
      operationId: logout
      parameters:
        - $ref: '#/components/parameters/XProfileId'
      responses:
        '204':
          description: Logged out successfully

  /me:
    get:
      tags: [Auth]
      summary: Get current selected profile context
      operationId: getCurrentUser
      parameters:
        - $ref: '#/components/parameters/XProfileId'
      responses:
        '200':
          description: Current user context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserContext'
        '404':
          $ref: '#/components/responses/NotFound'

  /chat/sessions:
    get:
      tags: [Chat]
      summary: List chat sessions for the selected profile
      operationId: listChatSessions
      parameters:
        - $ref: '#/components/parameters/XProfileId'
      responses:
        '200':
          description: Session list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatSessionSummary'
    post:
      tags: [Chat]
      summary: Create a new chat session
      operationId: createChatSession
      parameters:
        - $ref: '#/components/parameters/XProfileId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatSessionRequest'
      responses:
        '201':
          description: Created session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionDetail'

  /chat/sessions/{sessionId}:
    get:
      tags: [Chat]
      summary: Get one session snapshot
      operationId: getChatSession
      parameters:
        - $ref: '#/components/parameters/XProfileId'
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Chat]
      summary: Rename a chat session
      operationId: renameChatSession
      parameters:
        - $ref: '#/components/parameters/XProfileId'
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameChatSessionRequest'
      responses:
        '200':
          description: Updated session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSessionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Chat]
      summary: Delete a chat session and its messages/draft
      operationId: deleteChatSession
      parameters:
        - $ref: '#/components/parameters/XProfileId'
        - $ref: '#/components/parameters/SessionId'
      responses:
        '204':
          description: Session deleted
        '404':
          $ref: '#/components/responses/NotFound'

  /chat/sessions/{sessionId}/messages:
    get:
      tags: [Chat]
      summary: List messages for a session
      operationId: listChatMessages
      parameters:
        - $ref: '#/components/parameters/XProfileId'
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Ordered messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
    post:
      tags: [Chat]
      summary: Send a user text message
      operationId: sendChatMessage
      parameters:
        - $ref: '#/components/parameters/XProfileId'
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: |
            Final turn response as JSON, or streamed events if the client asks for
            `text/event-stream`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatTurnResponse'
            text/event-stream:
              schema:
                type: string
                example: |
                  event: user-message
                  data: {"message":{"messageId":"01H...","role":"USER","kind":"TEXT","text":"Pay Bob 500 HKD today"}}

                  event: assistant-message-start
                  data: {"messageId":"01H..."}

                  event: assistant-message-delta
                  data: {"delta":"Please confirm Bob"}

                  event: assistant-message-complete
                  data: {"session":{"sessionId":"01H...","state":"AWAITING_CONFIRMATION"}}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /chat/sessions/{sessionId}/events:
    post:
      tags: [Chat]
      summary: Submit a structured UI event such as payee selection or confirm click
      operationId: submitUiEvent
      parameters:
        - $ref: '#/components/parameters/XProfileId'
        - $ref: '#/components/parameters/SessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UiEventRequest'
      responses:
        '200':
          description: |
            Final turn response as JSON, or streamed events if the client asks for
            `text/event-stream`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatTurnResponse'
            text/event-stream:
              schema:
                type: string
                example: |
                  event: assistant-message-start
                  data: {"messageId":"01H..."}

                  event: assistant-message-delta
                  data: {"delta":"Payment submitted successfully."}

                  event: assistant-message-complete
                  data: {"session":{"sessionId":"01H...","state":"COMPLETED"}}
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    XProfileId:
      in: header
      name: X-Profile-Id
      required: true
      description: Selected demo profile identifier
      schema:
        type: string
        example: 01HX8QYP8W4G9Y0S2AY2J67NZB

    SessionId:
      in: path
      name: sessionId
      required: true
      schema:
        type: string
        example: 01HX8R5A1NQY39XK0D28S5B3EW

  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ProfileSummary:
      type: object
      required:
        - id
        - guid
        - permNetId
        - code
        - username
        - displayName
        - locale
        - status
        - supportedCapabilities
      properties:
        id:
          type: string
          example: 01HX8QYP8W4G9Y0S2AY2J67NZB
        guid:
          type: string
          nullable: true
          example: profile-poc-guid
        permNetId:
          type: string
          nullable: true
          example: 11114418_O88
        code:
          type: string
          example: HK_STAFF_001
        username:
          type: string
          description: Downstream identity value used by the backend.
          example: payment10
        displayName:
          type: string
          example: Victor Zhong
        avatarUrl:
          type: string
          nullable: true
          example: /assets/avatar.png
        locale:
          type: string
          example: en-HK
        status:
          $ref: '#/components/schemas/ProfileStatus'
        supportedCapabilities:
          type: array
          items:
            $ref: '#/components/schemas/CapabilityType'
          example: [REGISTERED_PAYEE_LOOKUP, DOMESTIC_PAYMENT]

    ProfileLoginRequest:
      type: object
      required:
        - profileId
        - password
      properties:
        profileId:
          type: string
          example: 01HX8QYP8W4G9Y0S2AY2J67NZB
        password:
          type: string
          example: tb123

    CurrentUserContext:
      type: object
      required:
        - profileId
        - username
        - displayName
        - locale
        - loginMode
        - supportedCapabilities
      properties:
        profileId:
          type: string
        username:
          type: string
          example: payment10
        displayName:
          type: string
          example: Victor Zhong
        avatarUrl:
          type: string
          nullable: true
        locale:
          type: string
          example: en-HK
        loginMode:
          type: string
          enum: [PROFILE_SELECTION]
        supportedCapabilities:
          type: array
          items:
            $ref: '#/components/schemas/CapabilityType'
          example: [REGISTERED_PAYEE_LOOKUP, DOMESTIC_PAYMENT]

    CreateChatSessionRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 120
          example: Pay Bob

    RenameChatSessionRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
          example: Pay Bob Rent

    ChatSessionSummary:
      type: object
      required:
        - sessionId
        - title
        - titleLocked
        - status
        - state
        - createdAt
        - updatedAt
      properties:
        sessionId:
          type: string
        title:
          type: string
          example: Pay Bob
        titleLocked:
          type: boolean
          description: True when the user manually renamed the session.
        status:
          $ref: '#/components/schemas/ChatSessionStatus'
        state:
          $ref: '#/components/schemas/ConversationState'
        llmProvider:
          allOf:
            - $ref: '#/components/schemas/LlmProviderType'
          nullable: true
        lastMessagePreview:
          type: string
          nullable: true
          example: Please confirm the payee and amount.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ChatSessionDetail:
      type: object
      required:
        - sessionId
        - title
        - titleLocked
        - status
        - state
        - createdAt
        - updatedAt
      properties:
        sessionId:
          type: string
        title:
          type: string
        titleLocked:
          type: boolean
          description: True when the user manually renamed the session.
        status:
          $ref: '#/components/schemas/ChatSessionStatus'
        state:
          $ref: '#/components/schemas/ConversationState'
        llmProvider:
          allOf:
            - $ref: '#/components/schemas/LlmProviderType'
          nullable: true
        activeDraft:
          allOf:
            - $ref: '#/components/schemas/PaymentDraft'
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ChatMessage:
      type: object
      required:
        - messageId
        - sessionId
        - role
        - kind
        - createdAt
      properties:
        messageId:
          type: string
        sessionId:
          type: string
        role:
          $ref: '#/components/schemas/MessageRole'
        kind:
          $ref: '#/components/schemas/MessageKind'
        text:
          type: string
          nullable: true
        contentBlocks:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ContentBlock'
        metadata:
          type: object
          nullable: true
          description: Optional diagnostics such as assistant `processingMs`.
          additionalProperties: true
        createdAt:
          type: string
          format: date-time

    SendMessageRequest:
      type: object
      required:
        - messageText
      properties:
        messageText:
          type: string
          minLength: 1
          maxLength: 4000
          example: Pay Bob 500 HKD today
        clientMessageId:
          type: string
          nullable: true
          example: cli-msg-001
        stream:
          type: boolean
          default: false
          description: If true, the client expects an event-stream response.

    UiEventRequest:
      type: object
      required:
        - eventType
        - sourceMessageId
        - sourceBlockId
      properties:
        eventType:
          $ref: '#/components/schemas/UiEventType'
        sourceMessageId:
          type: string
        sourceBlockId:
          type: string
        selectedItemId:
          type: string
          nullable: true
        actionValue:
          type: string
          nullable: true
        formValues:
          type: object
          nullable: true
          additionalProperties:
            type: string
        stream:
          type: boolean
          default: false

    ChatTurnResponse:
      type: object
      required:
        - session
        - assistantMessage
        - serverTimestamp
      properties:
        session:
          $ref: '#/components/schemas/ChatSessionDetail'
        userMessage:
          allOf:
            - $ref: '#/components/schemas/ChatMessage'
          nullable: true
        assistantMessage:
          $ref: '#/components/schemas/ChatMessage'
        activeDraft:
          allOf:
            - $ref: '#/components/schemas/PaymentDraft'
          nullable: true
        serverTimestamp:
          type: string
          format: date-time

    PaymentDraft:
      type: object
      required:
        - draftId
        - status
        - paymentType
        - lastUpdatedAt
      properties:
        draftId:
          type: string
        status:
          $ref: '#/components/schemas/PaymentDraftStatus'
        paymentType:
          $ref: '#/components/schemas/PaymentType'
        payeeQueryText:
          type: string
          nullable: true
        selectedPayee:
          allOf:
            - $ref: '#/components/schemas/PayeeSummary'
          nullable: true
        amount:
          type: string
          format: decimal
          nullable: true
          description: Decimal string serialized by the backend to avoid precision loss.
          example: "500.00"
        currency:
          type: string
          nullable: true
          example: HKD
        paymentDate:
          type: string
          format: date
          nullable: true
        downstreamReference:
          type: string
          nullable: true
        lastError:
          allOf:
            - $ref: '#/components/schemas/ErrorSummary'
          nullable: true
        context:
          type: object
          nullable: true
          additionalProperties: true
        lastUpdatedAt:
          type: string
          format: date-time

    PayeeSummary:
      type: object
      required:
        - payeeId
        - name
      properties:
        payeeId:
          type: string
          description: Opaque backend-managed payee identifier.
        name:
          type: string
        payeeType:
          type: string
          nullable: true
        bankCode:
          type: string
          nullable: true
        bankName:
          type: string
          nullable: true
        accountNumber:
          type: string
          nullable: true
        displayLabel:
          type: string
          nullable: true

    ContentBlock:
      oneOf:
        - $ref: '#/components/schemas/TextBlock'
        - $ref: '#/components/schemas/SummaryCardBlock'
        - $ref: '#/components/schemas/SelectableListBlock'
        - $ref: '#/components/schemas/InfoCardBlock'
        - $ref: '#/components/schemas/ErrorCardBlock'
      discriminator:
        propertyName: type
        mapping:
          TEXT: '#/components/schemas/TextBlock'
          SUMMARY_CARD: '#/components/schemas/SummaryCardBlock'
          SELECTABLE_LIST: '#/components/schemas/SelectableListBlock'
          INFO_CARD: '#/components/schemas/InfoCardBlock'
          ERROR_CARD: '#/components/schemas/ErrorCardBlock'

    TextBlock:
      type: object
      required:
        - blockId
        - type
        - text
      properties:
        blockId:
          type: string
        type:
          type: string
          enum: [TEXT]
        title:
          type: string
          nullable: true
        text:
          type: string
        metadata:
          type: object
          additionalProperties: true

    SummaryCardBlock:
      type: object
      required:
        - blockId
        - type
        - title
        - fields
      properties:
        blockId:
          type: string
        type:
          type: string
          enum: [SUMMARY_CARD]
        title:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/DisplayField'
        metadata:
          type: object
          additionalProperties: true

    SelectableListBlock:
      type: object
      required:
        - blockId
        - type
        - title
        - items
      properties:
        blockId:
          type: string
        type:
          type: string
          enum: [SELECTABLE_LIST]
        title:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/SelectableItem'
        metadata:
          type: object
          additionalProperties: true

    InfoCardBlock:
      type: object
      required:
        - blockId
        - type
        - title
        - text
      properties:
        blockId:
          type: string
        type:
          type: string
          enum: [INFO_CARD]
        title:
          type: string
        text:
          type: string
        metadata:
          type: object
          additionalProperties: true

    ErrorCardBlock:
      type: object
      required:
        - blockId
        - type
        - title
        - text
      properties:
        blockId:
          type: string
        type:
          type: string
          enum: [ERROR_CARD]
        title:
          type: string
        text:
          type: string
        metadata:
          type: object
          additionalProperties: true

    SelectableItem:
      type: object
      required:
        - itemId
        - label
      properties:
        itemId:
          type: string
        label:
          type: string
        description:
          type: string
          nullable: true
        value:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true

    DisplayField:
      type: object
      required:
        - label
        - value
      properties:
        label:
          type: string
        value:
          type: string

    ErrorSummary:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string

    ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          example: CHAT2PAY-400
        message:
          type: string
          example: Invalid request payload.
        details:
          type: array
          items:
            type: string
        timestamp:
          type: string
          format: date-time

    ProfileStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE

    CapabilityType:
      type: string
      enum:
        - REGISTERED_PAYEE_LOOKUP
        - DOMESTIC_PAYMENT
        - INTERNATIONAL_PAYMENT

    ChatSessionStatus:
      type: string
      enum:
        - ACTIVE
        - COMPLETED
        - FAILED
        - CANCELLED
        - ARCHIVED

    ConversationState:
      type: string
      enum:
        - IDLE
        - COLLECTING_DETAILS
        - AWAITING_PAYEE_SELECTION
        - AWAITING_CONFIRMATION
        - EXECUTING
        - COMPLETED
        - FAILED
        - CANCELLED

    PaymentDraftStatus:
      type: string
      enum:
        - DRAFT
        - AWAITING_CONFIRMATION
        - EXECUTING
        - CONFIRMED
        - FAILED
        - CANCELLED

    PaymentType:
      type: string
      enum:
        - DOMESTIC_PAYMENT
        - INTERNATIONAL_PAYMENT

    LlmProviderType:
      type: string
      enum:
        - COPILOT_PERSONAL
        - REMOTE_API

    MessageRole:
      type: string
      enum:
        - USER
        - ASSISTANT
        - SYSTEM

    MessageKind:
      type: string
      enum:
        - TEXT
        - BLOCKS
        - UI_EVENT
        - SYSTEM

    UiEventType:
      type: string
      enum:
        - SELECT_ITEM
        - CLICK_ACTION
        - SUBMIT_FORM